De-flake the overwrite-guard batch test and share metric helpers
The batch-overwrite test waits for the broker buffer to hold both
records so split claims cannot misroute the assertion between the
batch and stored detection branches. The Prometheus scrape helpers
move into tests/test_helpers.rs, replacing three per-file copies, and
the dispatch guard logs read/decode failures at debug so a
persistently blind detector stays observable.
