+++ src/job_store_shard/dequeue.rs
@@ -1,5 +1,7 @@
 //! Task dequeue and processing operations.
 
+use std::collections::HashSet;
+
 use slatedb::WriteBatch;
 use slatedb::config::WriteOptions;
 
@@ -48,7 +50,7 @@ struct DequeueIterationState {
     /// batch is uncommitted, so a `db.get` cannot see those writes; the
     /// double-dispatch detection guard consults this set to catch a repeat
     /// of the same task id within one dequeue call.
-    leased_task_ids: std::collections::HashSet<String>,
+    leased_task_ids: HashSet<String>,
     processed_internal: bool,
 }
 
@@ -64,7 +66,7 @@ impl DequeueIterationState {
             pending_attempts: Vec::new(),
             holder_releases: Vec::new(),
             converted_requests: Vec::new(),
-            leased_task_ids: std::collections::HashSet::new(),
+            leased_task_ids: HashSet::new(),
             processed_internal: false,
         }
     }
@@ -1204,28 +1206,46 @@ impl JobStoreShard {
                     crate::metrics::LeaseOverwriteSource::Batch,
                 );
             }
-        } else if let Ok(Some(existing_bytes)) = self.db.get(&leased_task_key(task_id)).await
-            && let Ok(existing) = crate::codec::decode_lease(existing_bytes)
-        {
-            let remaining_lease_ms = existing.expiry_ms() - now_ms;
-            if remaining_lease_ms > 0 {
-                tracing::warn!(
-                    tenant = %tenant,
-                    job_id = %job_id,
+        } else {
+            // Detection must never alter dispatch: read and decode failures
+            // are logged at debug so a persistently blind detector remains
+            // observable, then dispatch proceeds regardless.
+            match self.db.get(&leased_task_key(task_id)).await {
+                Ok(Some(existing_bytes)) => match crate::codec::decode_lease(existing_bytes) {
+                    Ok(existing) => {
+                        let remaining_lease_ms = existing.expiry_ms() - now_ms;
+                        if remaining_lease_ms > 0 {
+                            tracing::warn!(
+                                tenant = %tenant,
+                                job_id = %job_id,
+                                task_id = %task_id,
+                                existing_worker_id = %existing.worker_id(),
+                                new_worker_id = %worker_id,
+                                remaining_lease_ms,
+                                "dispatching RunAttempt over a still-live existing lease; \
+                                 possible double dispatch",
+                            );
+                            if let Some(m) = &self.metrics {
+                                m.record_task_lease_overwrite(
+                                    self.name(),
+                                    decoded.task_group(),
+                                    crate::metrics::LeaseOverwriteSource::Stored,
+                                );
+                            }
+                        }
+                    }
+                    Err(e) => tracing::debug!(
+                        task_id = %task_id,
+                        error = %e,
+                        "lease-overwrite guard could not decode existing lease record",
+                    ),
+                },
+                Ok(None) => {}
+                Err(e) => tracing::debug!(
                     task_id = %task_id,
-                    existing_worker_id = %existing.worker_id(),
-                    new_worker_id = %worker_id,
-                    remaining_lease_ms,
-                    "dispatching RunAttempt over a still-live existing lease; \
-                     possible double dispatch",
-                );
-                if let Some(m) = &self.metrics {
-                    m.record_task_lease_overwrite(
-                        self.name(),
-                        decoded.task_group(),
-                        crate::metrics::LeaseOverwriteSource::Stored,
-                    );
-                }
+                    error = %e,
+                    "lease-overwrite guard read failed",
+                ),
             }
         }
         state.leased_task_ids.insert(task_id.to_string());
