+++ tests/task_scanner_saturation_tests.rs
@@ -474,24 +474,3 @@ async fn scanner_buffers_ready_then_breaks_on_future() {
         );
     });
 }
-
-fn gather_metrics_text(metrics: &silo::metrics::Metrics) -> String {
-    use prometheus::{Encoder, TextEncoder};
-    let encoder = TextEncoder::new();
-    let metric_families = metrics.registry().gather();
-    let mut buffer = Vec::new();
-    encoder.encode(&metric_families, &mut buffer).unwrap();
-    String::from_utf8(buffer).unwrap()
-}
-
-fn extract_metric_value(body: &str, substrings: &[&str]) -> f64 {
-    let line = body
-        .lines()
-        .find(|l| !l.starts_with('#') && substrings.iter().all(|s| l.contains(s)))
-        .unwrap_or_else(|| panic!("metric line not found for substrings {:?}", substrings));
-    line.rsplit_once(' ')
-        .unwrap_or_else(|| panic!("no space-separated value in line: {}", line))
-        .1
-        .parse::<f64>()
-        .unwrap_or_else(|_| panic!("could not parse metric value from line: {}", line))
-}
